requirements.txt
requests
websockets
fastapi
uvicorn
aioredis
psycopg2
asyncpg
sqlalchemy
databases
alembic
python-dotenv
logging